requirements.txt
google-generativeai>=0.3.0
watchdog>=3.0.0
openai>=1.0.0
anthropic>=0.75.0
markdown2>=2.5.0

